Mod: add Chunk Loaders mod
This is to make the Advanced Wireless Terminal from Tom's Simple Storage
work at any distance and across dimensions using a level 4 beacon. Tom's
mod doesn't work unless the beacon and storage system are in a loaded
chunk, so we need to chunk load that spot in order to access it from
anywhere with the wireless terminal.